% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_theme.R
\name{bicycle_cols}
\alias{bicycle_cols}
\title{Function to extract bicycle colors as hex codes}
\usage{
bicycle_cols(...)
}
\arguments{
\item{...}{Character names of bicycle_colors}
}
\value{
a vector of hex codes
}
\description{
Function to extract bicycle colors as hex codes
}
\examples{
bicycle_cols()
bicycle_cols('indigo')
bicycle_cols('forestgreen', 'lockwood')
}
