% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_theme.R
\name{scale_color_bicycle}
\alias{scale_color_bicycle}
\title{Color scale constructor for bicycle colors}
\usage{
scale_color_bicycle(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in bicycle_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for bicycle colors
}
\examples{
ggplot(mtcars, aes(wt, mpg, color = mpg)) +
geom_point(size = 4) + theme_bicycle() +
 scale_color_bicycle(palette = 'main', discrete = FALSE)
 ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) + theme_bicycle() +
  scale_color_bicycle(palette = 'main')
}
