% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_theme.R
\name{scale_fill_bicycle}
\alias{scale_fill_bicycle}
\title{Fill scale constructor for bicycle colors}
\usage{
scale_fill_bicycle(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in bicycle_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for bicycle colors
}
\examples{
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
geom_bar() +
  theme_bicycle() +
  scale_fill_bicycle(palette = "grays")
}
