% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_theme.R
\name{theme_bicycle}
\alias{theme_bicycle}
\title{Bicycle theme}
\usage{
theme_bicycle(...)
}
\arguments{
\item{...}{Any additional ggplot2 theme arguments that need to be overwritten}
}
\value{
custom theme
}
\description{
Bicycle theme
}
\examples{
theme_bicycle()
}
